requirements.txt
requests
python-dotenv
hyperliquid-python-sdk
python-telegram-bot
